[update] ステータス関連スクリプトを同じフォルダにまとめた
